@@ -162,7 +162,7 @@ class _ManifestPageState extends State<ManifestPage> {
 
   void refreshToken() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    var response = await http.get(StringData.refreshTokenAPI,
+    var response = await http.get(Uri.parse(StringData.refreshTokenAPI),
         headers: {'x-access-tokens': prefs.getString('token')});
     var result = json.decode(response.body);
     if (result['result'] == 'verified')
